Updates/fixes related to remote/proxy services

+++ install_remote.sh
@@ -3,9 +3,14 @@ set -e
 
 HERE_M3U8="$(cd "$(dirname "${BASH_SOURCE[0]}")" >/dev/null 2>&1 && pwd)"
 source "$HERE_M3U8"/prep_env.sh
+UPDATE_ONLY_NON_FRPC=false
+if [ "$1" == 'updateNonProxy' ]; then
+    UPDATE_ONLY_NON_FRPC=true
+fi
 
-printTitle "Install Docker"
-ssh -A -t "$PROXY_SSH_STRING" bash -l <<-EOF
+if [ "$UPDATE_ONLY_NON_FRPC" != true ]; then
+    printTitle "Install Docker"
+    ssh -A -t "$PROXY_SSH_STRING" bash -l <<-EOF
 sudo install -m 0755 -d /etc/apt/keyrings
 curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo gpg --dearmor -o /etc/apt/keyrings/docker.gpg
 sudo chmod a+r /etc/apt/keyrings/docker.gpg
@@ -20,20 +25,31 @@ sudo apt-get update -y
 sudo apt-get install -y docker-ce docker-ce-cli containerd.io docker-buildx-plugin docker-compose-plugin
 sudo usermod -aG docker \$USER
 EOF
-echo "Done."
+    echo "Done."
 
-printTitle "Prepare FRPS dependencies"
-syncRemoteEnvFileIfUndefined "$PROXY_SSH_STRING" "$PROXY_HOME/landscape-remote-services/state/frps-tokens.txt" "$MAIN_NODE_HOSTNAME_LOWERCASE" "$(echo $RANDOM | sha512sum | awk '{print $1}')$(echo $RANDOM | sha512sum | awk '{print $1}')" "$HERE_M3U8"/files/frps-tokens.txt
-scp -q "$HERE_M3U8"/files/frps-tokens.txt "$PROXY_SSH_STRING":~/landscape-remote-services/state/frps-tokens.txt
-scp -q "$HERE_M3U8"/files/openTCPPort.sh "$PROXY_SSH_STRING":~/landscape-remote-services/openTCPPort.sh
-rm "$HERE_M3U8"/files/frps-tokens.txt
-scp "$HERE_M3U8"/files/frps.create-image.sh "$PROXY_SSH_STRING":~/landscape-remote-services/frps.create-image.sh
-ssh -A -t "$PROXY_SSH_STRING" "bash '$PROXY_HOME/landscape-remote-services/frps.create-image.sh'"
-echo "Done."
+    printTitle "Prepare FRPS dependencies"
+    syncRemoteEnvFileIfUndefined "$PROXY_SSH_STRING" "$PROXY_HOME/landscape-remote-services/state/frps-tokens.txt" "$MAIN_NODE_HOSTNAME_LOWERCASE" "$(echo $RANDOM | sha512sum | awk '{print $1}')$(echo $RANDOM | sha512sum | awk '{print $1}')" "$HERE_M3U8"/files/frps-tokens.txt
+    scp -q "$HERE_M3U8"/files/frps-tokens.txt "$PROXY_SSH_STRING":~/landscape-remote-services/state/frps-tokens.txt
+    scp -q "$HERE_M3U8"/files/openTCPPort.sh "$PROXY_SSH_STRING":~/landscape-remote-services/openTCPPort.sh
+    rm "$HERE_M3U8"/files/frps-tokens.txt
+    bash "$HERE_M3U8"/files/frps.create-image.sh
+    docker image save imranrdev/frps-with-multiuser:latest -o /tmp/frps-with-multiuser.tar
+    docker rmi imranrdev/frps-with-multiuser:latest 2>/dev/null || :
+    scp /tmp/frps-with-multiuser.tar "$PROXY_SSH_STRING":~/landscape-remote-services/frps-with-multiuser.tar
+    rm /tmp/frps-with-multiuser.tar
+    ssh -A -t "$PROXY_SSH_STRING" "docker rmi imranrdev/frps-with-multiuser:latest 2>/dev/null || :"
+    ssh -A -t "$PROXY_SSH_STRING" "docker image load -i '$PROXY_HOME/landscape-remote-services/frps-with-multiuser.tar'"
+    echo "Done."
+fi
 
 printTitle "Prepare strelaysrv dependencies"
-scp "$HERE_M3U8"/files/strelaysrv.create-image.sh "$PROXY_SSH_STRING":~/landscape-remote-services/strelaysrv.create-image.sh
-ssh -A -t "$PROXY_SSH_STRING" "bash '$PROXY_HOME/landscape-remote-services/strelaysrv.create-image.sh'"
+bash "$HERE_M3U8"/files/strelaysrv.create-image.sh
+docker image save imranrdev/strelaysrv-docker:latest -o /tmp/strelaysrv-docker.tar
+docker rmi imranrdev/strelaysrv-docker:latest 2>/dev/null || :
+scp /tmp/strelaysrv-docker.tar "$PROXY_SSH_STRING":~/landscape-remote-services/strelaysrv-docker.tar
+rm /tmp/strelaysrv-docker.tar
+ssh -A -t "$PROXY_SSH_STRING" "docker rmi imranrdev/strelaysrv-docker:latest 2>/dev/null || :"
+ssh -A -t "$PROXY_SSH_STRING" "docker image load -i '$PROXY_HOME/landscape-remote-services/strelaysrv-docker.tar'"
 echo "Done."
 
 printTitle "Prepare Logtfy dependencies"
@@ -44,6 +60,7 @@ jq '.moduleCustomization |= map(select(.module == "ssh_logins" or .module == "po
 scp -q "$HERE_M3U8"/files/logtfy.remote.json "$PROXY_SSH_STRING":~/landscape-remote-services/state/logtfy.json
 rm "$HERE_M3U8"/files/logtfy.remote.json
 rm "$HERE_M3U8"/files/logtfy.remote.temp.json
+docker pull imranrdev/logtfy
 echo "Done."
 
 printTitle "Generate Docker Compose and Systemd Files and Start the Service"
@@ -54,12 +71,19 @@ scp "$HERE_M3U8"/files/landscape-remote.service "$PROXY_SSH_STRING":~/landscape-
 scp "$HERE_M3U8"/files/landscape-remote.docker-compose.yaml "$PROXY_SSH_STRING":~/landscape-remote-services/state/landscape-remote.docker-compose.yaml
 rm "$HERE_M3U8"/files/landscape-remote.docker-compose.yaml
 rm "$HERE_M3U8"/files/landscape-remote.service
-ssh -A -t "$PROXY_SSH_STRING" "bash '$PROXY_HOME/landscape-remote-services/landscape-remote.install.sh'"
+if [ "$UPDATE_ONLY_NON_FRPC" != true ]; then
+    ssh -A -t "$PROXY_SSH_STRING" "bash '$PROXY_HOME/landscape-remote-services/landscape-remote.install.sh'"
+else
+    ssh -A -t "$PROXY_SSH_STRING" "bash '$PROXY_HOME/landscape-remote-services/landscape-remote.install.sh' logtfy"
+    ssh -A -t "$PROXY_SSH_STRING" "bash '$PROXY_HOME/landscape-remote-services/landscape-remote.install.sh' strelaysrv"
+fi
 echo "Done."
 
-if "$HERE_M3U8"/files/check_root_luks.sh >/dev/null 2>&1; then
-    printTitle "Install FRPC-Preboot and Dracut-Crypt-SSH so that root volume can be decrypted remotely."
-    bash "$HERE_M3U8"/files/dracut-crypt-ssh.install.sh
-    bash "$HERE_M3U8"/files/frpc-preboot.install.sh
-    rm "$HERE_M3U8"/files/frpc-preboot.ini
+if [ "$UPDATE_ONLY_NON_FRPC" != true ]; then
+    if "$HERE_M3U8"/files/check_root_luks.sh >/dev/null 2>&1; then
+        printTitle "Install FRPC-Preboot and Dracut-Crypt-SSH so that root volume can be decrypted remotely."
+        bash "$HERE_M3U8"/files/dracut-crypt-ssh.install.sh
+        bash "$HERE_M3U8"/files/frpc-preboot.install.sh
+        rm "$HERE_M3U8"/files/frpc-preboot.ini
+    fi
 fi
